fix(cleanUp): removes all created records in reverse order
This helps to overcome foreign key constraints issue in SQL based databases. When you create a record first the deepest association is created and only then the record which you want. So, in order to satisfy foreign key we need to start removing objects from the top most record, record which is not referenced by any other records.

Fixes #110